Refactor/Cleanup Initial Code
The initial code was refactored into three different types:
the `FuncProcessor`, which creates argspecs and extracts raw
documentation, the `CodeGenerator`, whose subclasses generate
the actual output code, and the `CodeLookup`, whose subclasses
are used by the `FuncProcessor` to look up the appropriate code
for transformers, etc.  The `CodeGenerator` and `CodeLookup`
together form a single language backend.