Support 'optional' output arg tag
Output arg tags are the prefix before the type name in the type
annotation in output argspecs.  Previously, only the 'nullable'
tag was supported.  Multiple tags may be used by separating them
with a ', '.

The 'optional' output arg tag specifies that a particular output
argument is only needed if the corresponding function parameter
is set to True.  If the corresponding function parameter is False,
a null pointer is passed to the C function, and None is returned
for that output arg.  Otherwise, a pointer to the type of the output
arg is automatically constructed, passed to the C function, and
retrieved and processed for output.